Move _tus_chunk_action to view_utils
